Finally

+++ FactoryMethodPattern/CoffeeFactory.java
@@ -1,7 +1,7 @@
 package FactoryMethodPattern;
 
 public class CoffeeFactory {
-    public static Coffee createCoffee(String type, String name, int price) {
+    public static Coffee createCoffee(String type, String name, double price) {
         if (type.equalsIgnoreCase("hot")) {
             return new HotCoffee(name, price);
         } else if (type.equalsIgnoreCase("ice")) {
